backend/requirements.txt
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
openai==1.76.2
pillow==11.2.1
dotenv==0.9.9
drf-spectacular==0.28.0
adrf==0.1.9
aiohttp==3.11.18
requests
psycopg2-binary
celery
redis
django-celery-beat
celery-once